Adds more to the tests as well as cleans them up and splits them to their own individual tests for validation.
Removes standalone class for the no operation EnergyBuffer instance. It is `EnergyHandlerUtil.EMPTY`

Adds a temporary deprecation path for IEnergyStorage to be IEnergyHandlers, but this might be removed. This also redirects all `Capabilities.EnergyStorage` to `Capabilities.EnergyHandler`

Adds more documentation to communicate more clearly what should be used

Adds a max value of 2 billion instead of 2.147.. billion. This serves as a more readable cap without really sacrificing the functionality.

Ran Spotless